Refactor:에러처리 파일 형식 동일화 및 error 폴더 생성
