+++ __tests__/LottoTest.js
@@ -1,4 +1,4 @@
-import Lotto from "../src/Lotto.js";
+import Lotto from "../src/errors/Lotto.js";
 
 describe("로또 클래스 테스트", () => {
   const spaceTestCases = [
